% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{all_missing_dich}
\alias{all_missing_dich}
\title{all_missing_dich function}
\usage{
all_missing_dich(
  vars,
  data,
  allmissyes = 1,
  allmissno = 0,
  sumisna_fun = "is.na"
)
}
\arguments{
\item{vars}{vector of variables to evaluate}

\item{data}{dataset containing the variables}

\item{allmissyes}{what to return if all of the variables are NA. Default is 1}

\item{allmissno}{what to return if all of the variables are not NA. Defalt is 0}

\item{sumisna_fun}{what function to pass to the sumisna function. default is 'is.na'}
}
\description{
given a vector of variable names and a dataset this function will find all of the variables in the vector that are present in the dataset and then of those determine (using sumisna) if all of them are NA.
}
\examples{
all_missing_dich_function()

}
\keyword{all}
\keyword{all_missing_dich}
\keyword{any}
\keyword{dich}
\keyword{missing}
\keyword{missingness}
\keyword{sumisna}
