% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{ASED}
\alias{ASED}
\title{ASED function}
\usage{
ASED(x = x, y = y, method = "euclidean")
}
\arguments{
\item{x}{case 1 (must have the same number of variables as y)}

\item{y}{case 2 (must have the same number of variables as x)}

\item{method}{the method to be used in the ASED function. default is 'eucludean'}
}
\description{
computes average euclidean or squared euclidean distance
}
\examples{
ASED_function()
}
\keyword{ASED}
\keyword{cluster}
\keyword{distance}
\keyword{inter-cluster}
\keyword{outlier}
