% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{ci.dif}
\alias{ci.dif}
\title{ci.dif function}
\usage{
ci.dif(x, n, rownm = "categories", t100 = TRUE)
}
\arguments{
\item{x}{a vector of frequencies of multiple categories for which a binomial confidence interval should be created}

\item{n}{number of cases used to create the frequencies}

\item{rownm}{name of the first varaible in the returned dataframe (name of the frequency categories). defaults to 'categories'}

\item{t100}{should the values be multiplied by 100 (used if they are percentages/proportions). defaults to TRUE}
}
\description{
creates a matrix of
}
\examples{
ci.dif_function()
}
\keyword{ci.dif}
\keyword{confidence}
\keyword{interval}
