% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{clear.labels}
\alias{clear.labels}
\title{clear labels function}
\usage{
clear.labels(x)
}
\arguments{
\item{x}{data frame to remove labels from}
}
\description{
if a dataset has labels in it then it will not join using dpylr, use this funciton to clear the labels out before merging
}
\examples{
clear.labels_function()
}
\keyword{clear}
\keyword{labels}
\keyword{remove}
