% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{clustrecs}
\alias{clustrecs}
\title{clustrecs function}
\usage{
clustrecs(indxdf)
}
\arguments{
\item{indxdf}{a dataframe of all indexes from clustIndex for which the best fit solution for each index will be calculated}
}
\description{
for calculating optimal cluster numbers based on clustIndex (from An Examination Of Indexes For Determining The Number Of Clusters In Binary Data Sets http://epub.wu.ac.at/1542/  xuindex from https://www.researchgate.net/figure/Example-of-the-Xu-index-applied-to-synthetic-data-Data-classified-into-5-classes-left_fig1_221472151)
}
\examples{
clustrecs_function()
}
\keyword{cluster}
\keyword{clustrecs}
\keyword{distance}
\keyword{fit}
\keyword{inter-cluster}
\keyword{outlier}
\keyword{solution}
