% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{collapse_list}
\alias{collapse_list}
\title{collapse list function}
\usage{
collapse_list(lst, delimeter = ", ")
}
\arguments{
\item{lst}{list to collapse}

\item{delimeter}{what is the lst delimited by. default is ", "}
}
\description{
collapses a list to a vector, useful for when a list is hiding in a dataframe and you need to convert it to a vector while collapsing vector elements like c("yes","yes,"no") to atomic elements like "yes, yes, no"
}
\examples{
collapse_list_function()
}
\keyword{collapse}
\keyword{list}
