% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{confint.geeglm}
\alias{confint.geeglm}
\title{confint.geeglm function}
\usage{
\method{confint}{geeglm}(object, parm, model_package = "geepack", level = 0.95, ...)
}
\arguments{
\item{object}{gee model for which to create confidence intervals}

\item{parm}{not sure but it came with the function}

\item{level}{level of confidence needed. defaults to '.95'}

\item{...}{extra arguments}
}
\description{
confidence intervals for generalized estimating equation models as found here https://stackoverflow.com/questions/21221280/confidence-interval-of-coefficients-using-generalized-estimating-equation-gee
}
\examples{
confint.geeglm_function()
}
\keyword{chisquare}
\keyword{confint.geeglm}
\keyword{x2}
