% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{create_studyid}
\alias{create_studyid}
\title{create_studyid function}
\usage{
create_studyid(x, sha = FALSE)
}
\arguments{
\item{x}{a seed unique to a patient used for generating the pseudorandom id.}
}
\description{
given a seed this function creates a pseudo random ID that should be unique to a patient. the seed also must be unique to the patient but preferably not a single identifier. for example it would be better to use ssn multiplied by dob and gender rather than ssn alone. additionally row numbers can be used if they are unlikely to yield patient information
}
\examples{
create_studyid_function()

}
\keyword{create_studyid}
\keyword{de-identify}
\keyword{deidentify}
\keyword{id}
\keyword{study}
\keyword{studyid}
