% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enr functions.R
\name{cut_performance}
\alias{cut_performance}
\title{cut_performance function}
\usage{
cut_performance(predicted_probs, cutpoints, fit_met, y_true)
}
\arguments{
\item{predicted_probs}{vector of predicted probabilities coming from a logistic model}

\item{cutpoints}{vector of cut points to loop through to and produce classification based on predicted probabilities}

\item{fit_met}{the fit metric used to evaluate the classification given the cutoff}

\item{y_true}{vector of 'true' values to be compared against classification based on predicted probs and cutoffs}
}
\description{
fucntion to calculate optimal cut performance (DOCUMENTATION COMING- CURRENT DOCUMENTATION INCORRECT)
}
\examples{
cut_performance()
}
\keyword{cut_performance}
\keyword{enr}
