% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{dfsmry}
\alias{dfsmry}
\title{dfsmry function}
\usage{
dfsmry(x, rowname = "Values")
}
\arguments{
\item{x}{the data frame you wish to summarize}

\item{rowname}{the name you would like to give the summary rows. defaults to 'Values'}
}
\description{
summarize a data frame in a table giving ncases, nvar, and average missing
}
\examples{
dfsmry_function()
}
\keyword{dataframe}
\keyword{dfsmry}
\keyword{frequency}
\keyword{summary}
\keyword{table}
