% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{dice}
\alias{dice}
\title{dice function}
\usage{
dice(x, y)
}
\arguments{
\item{x}{case 1 (the dataset columns in this case should contain only those used for matching)}

\item{y}{case 2 (the dataset columns in this case should contain only those used for matching)}
}
\description{
Dice coefficient (measure of agreement between two cases on suites of binary variables) (2(M.11)) / ((2(M.11)) + M.10 + M.01)
}
\examples{
dice_function()
}
\keyword{binary}
\keyword{dice}
\keyword{distance}
