% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R, R/highlandr.R
\name{dissm_dice}
\alias{dissm_dice}
\title{dissm_dice function}
\usage{
dissm_dice(y)

dissm_dice(y)
}
\arguments{
\item{y}{the dataset for which a pairwise dissimilarity matrix should be created (the dataset columns should contain only those used for clustering - i.e. the raw data no identifiers or other variables)}
}
\description{
returns a pairwise dissimilarity matrix using dice distance (i.e. 1-dice coefficient) to be used for dichotmous clustering

returns a pairwise dissimilarity matrix using dice distance (i.e. 1-dice coefficient) to be used for dichotmous clustering
}
\examples{
dissm_dice_function()
dissm_dice_function()
}
\keyword{binary}
\keyword{cluster}
\keyword{dice}
\keyword{dichotomous}
\keyword{dissm_dice}
\keyword{distance}
\keyword{matrix}
