% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R, R/highlandr.R
\name{dissm_jaccard}
\alias{dissm_jaccard}
\title{dissm_jaccard function}
\usage{
dissm_jaccard(y)

dissm_jaccard(y)
}
\arguments{
\item{y}{the dataset for which a pairwise dissimilarity matrix should be created (the dataset columns should contain only those used for clustering - i.e. the raw data no identifiers or other variables)}
}
\description{
returns a pairwise dissimilarity matrix using jaccard distance (i.e. 1-jaccard index) to be used for dichotmous clustering

returns a pairwise dissimilarity matrix using jaccard distance (i.e. 1-jaccard index) to be used for dichotmous clustering
}
\examples{
dissm_jaccard_function()
dissm_jaccard_function()
}
\keyword{binary}
\keyword{cluster}
\keyword{dichotomous}
\keyword{dissm_jaccard}
\keyword{distance}
\keyword{jaccard}
\keyword{matrix}
