% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R, R/highlandr.R
\name{dissm_smc}
\alias{dissm_smc}
\title{dissm_smc function}
\usage{
dissm_smc(y)

dissm_smc(y)
}
\arguments{
\item{y}{the dataset for which a pairwise dissimilarity matrix should be created (the dataset columns should contain only those used for clustering - i.e. the raw data no identifiers or other variables)}
}
\description{
returns a pairwise dissimilarity matrix using simple matching distance (i.e. 1-simple matching coefficient) to be used for dichotmous clustering

returns a pairwise dissimilarity matrix using simple matching distance (i.e. 1-simple matching coefficient) to be used for dichotmous clustering
}
\examples{
dissm_smc_function()
dissm_smc_function()
}
\keyword{binary}
\keyword{cluster}
\keyword{dichotomous}
\keyword{dissm_smc}
\keyword{distance}
\keyword{matching}
\keyword{matrix}
\keyword{simple}
\keyword{smc}
