% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R, R/highlandr.R
\name{dist_measure_explore}
\alias{dist_measure_explore}
\title{dist_measure_explore function}
\usage{
dist_measure_explore(
  data,
  cvars,
  mes = "rawdata",
  metric = "euclidean",
  method = "ward",
  diss = TRUE,
  maxclust = 15,
  bindata = FALSE
)

dist_measure_explore(
  data,
  cvars,
  mes = "rawdata",
  metric = "euclidean",
  method = "ward",
  diss = TRUE,
  maxclust = 15,
  bindata = FALSE
)
}
\arguments{
\item{data}{a dataframe containing only the identifier and  raw data to be clustered}

\item{cvars}{vector of names in order for all variables to be used in the clustering}

\item{mes}{vector of distance measures to use. default is distance measures from}

\item{metric}{metric to use in the agnes cluster analysis. default is 'euclidean'}

\item{method}{method of hierarchical clustering to use. default is 'ward'}

\item{diss}{whether dissimilarity matrix is being passed (TRUE) or raw data (FALSE). default is TRUE}

\item{maxclust}{maximum cluster solutions to produce. default is 15}

\item{bindata}{is the data being analyzed binary data (TRUE) or not (FALSE). default is FASLE}
}
\description{
massive function for exploratory hierachical cluster analysis varying distance measures,

massive function for exploratory hierachical cluster analysis varying distance measures,
}
\examples{
dist_measure_explore_function()
dist_measure_explore_function()
}
\keyword{cluster}
\keyword{dist_measure_explore}
\keyword{distance}
\keyword{fit}
\keyword{inter-cluster}
\keyword{outlier}
\keyword{solution}
