% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{distoon}
\alias{distoon}
\title{distoon function}
\usage{
distoon(x, meth)
}
\arguments{
\item{x}{the data used}

\item{meth}{the distance function (method) used}
}
\description{
allows a function (particularly a self defined function) to be used in the quoted call of another function. an example of this is the heatmaply function which takes an argument 'distfun' which expects a quoted distance function such as "spearman" which cannot be substituted for a regular function name. instead use distoon to set the name and call distoon in it's place in the functions argument
}
\examples{
distoon_function()
}
\keyword{distoon}
\keyword{function}
\keyword{proxy}
