% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enr functions.R
\name{en_kfold_accuracy_tied}
\alias{en_kfold_accuracy_tied}
\title{en_kfold_accuracy_tied function}
\usage{
en_kfold_accuracy_tied(
  tied_alphas,
  tied_cutoffs,
  ddata,
  response_var,
  iter = 10,
  k = 10,
  favor_parsimony = TRUE,
  eq_wt = FALSE,
  seed = 123,
  type_meas = "deviance"
)
}
\description{
This function breaks ties if model performance for enr models is tied (DOCUMENTATION COMING)
}
\examples{
en_kfold_accuracy_tied()

}
\keyword{en_kfold_accuracy_tied}
\keyword{enr}
\keyword{tied}
