% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enr functions.R
\name{en_kfold_accuracy_tied_grid}
\alias{en_kfold_accuracy_tied_grid}
\title{en_kfold_accuracy_tied_grid function}
\usage{
en_kfold_accuracy_tied_grid(
  tied_alphas,
  tied_lambdas,
  tied_cutoffs,
  ddata,
  response_var,
  iter = 100,
  k = 10,
  favor_parsimony = TRUE,
  eq_wt = FALSE,
  seed = 522,
  type_meas = "deviance",
  selection_type = "modal"
)
}
\description{
This function breaks ties if model performance for grid search enr models is tied (DOCUMENTATION COMING)
}
\examples{
en_kfold_accuracy_tied_grid()

}
\keyword{enr}
\keyword{grid}
\keyword{tied}
