% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{firstrow_to_vars}
\alias{firstrow_to_vars}
\title{firstrow_to_vars function}
\usage{
firstrow_to_vars(df, removefirstrow = TRUE, reset_rownames = TRUE)
}
\arguments{
\item{df}{a dataframe for which you want to turn the first row into variable names}

\item{removefirstrow}{boolean indicating whether to remove the first row after making it the df variable names. default is TRUE}

\item{reset_rownames}{booelan indicating whether to reset the row names after removing the first row. default is TRUE}
}
\description{
given a data frame takes the first row of the data frame and makes it the variable names and then removes the first row. this is useful when using rbind or when an imported dataset doesn't account for headers. duplicative with row_to_var but more optional and less error prone
}
\examples{
firstrow_to_vars_function()
}
\keyword{firstrow_to_vars}
\keyword{row}
\keyword{to}
\keyword{variable}
