% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{format_time_difference}
\alias{format_time_difference}
\title{format_time_difference function}
\usage{
format_time_difference(difftime_obj)
}
\arguments{
\item{difftime_obj}{A time difference object}
}
\description{
A function for formatting a time difference object into 'hours - minutes - seconds
}
\examples{
format_time_difference()
}
\keyword{format}
\keyword{progress}
\keyword{time}
\keyword{timer}
