% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{formula_maker}
\alias{formula_maker}
\title{formula_maker function}
\usage{
formula_maker(outcome, covars)
}
\arguments{
\item{outcome}{the outcome variable name. note, this should be in quotes (e.g. "outcome1")}

\item{covars}{the vector of covariates to collapse}
}
\description{
This function takes as it's input an outcome variable and a vector of covariates and creates a formula object with the covariates separated by '+'. Note this is only for linear models
}
\examples{
formula_maker_function()
}
\keyword{formula}
\keyword{formula_maker}
\keyword{linear}
\keyword{maker}
\keyword{model}
