% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{fp}
\alias{fp}
\title{fp function}
\usage{
fp(freq, perc)
}
\arguments{
\item{freq}{a value representing a frequency to be combined}

\item{perc}{a value representing a percent to be combined}
}
\description{
takes a frequency and percent variable and combines them in the form of 'x (xx%)'
}
\examples{
fp_function()
}
\keyword{fp}
\keyword{frequency}
\keyword{percent}
