% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{fp_msd_class}
\alias{fp_msd_class}
\title{fp_msd_class function}
\usage{
fp_msd_class(indvar, classvar, fp = TRUE, funct1 = "mean", funct2 = "sd")
}
\arguments{
\item{indvar}{independent variable to be split across the class variable. usually presented in the form data$variable}

\item{classvar}{class variable used to split the indvar. usually presented in the form data$variable}

\item{fp}{boolean indicating if the requested result should be frequency(percent). Default is TRUE}

\item{funct1}{the first function to use (i.e. outside the parentheses) if fp='FALSE'. default is 'mean'}

\item{funct2}{the second function to use (i.e. inside the parentheses) if fp='FALSE'. default is 'sd'}
}
\description{
base function for the fp_msd_class suite, recommend using fp_msd_class2. given an independent variable and a class variable creates a dataframe that is a cross tabulation of the two with values being 'out of' the class variable columns
}
\examples{
fp_msd_class_function()

}
\keyword{cross}
\keyword{crosstab}
\keyword{fp_msd_class}
\keyword{summary}
\keyword{variables}
