% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{fp_msd_class2}
\alias{fp_msd_class2}
\title{fp_msd_class2 function}
\usage{
fp_msd_class2(
  indvar,
  classvar = NULL,
  fp = TRUE,
  funct1 = "mean",
  funct2 = "sd",
  shownval = TRUE,
  total = TRUE,
  rnd_digs = 2,
  rownvar = "Class",
  count_miss = "ifmiss",
  count_miss_lab = "Missing"
)
}
\arguments{
\item{indvar}{independent variable to be split across the class variable. usually presented in the form data$variable}

\item{classvar}{class variable used to split the indvar. usually presented in the form data$variable. if left blank (i.e. no class) then the simple frequencies will be returned without any splitting}

\item{fp}{boolean indicating if the requested result should be frequency(percent). Default is TRUE}

\item{funct1}{the first function to use (i.e. outside the parentheses) if fp='FALSE'. default is 'mean'. if supplying different functions be sure to quote e.g. "IQR"}

\item{funct2}{the second function to use (i.e. inside the parentheses) if fp='FALSE'. default is 'sd'. if supplying different functions be sure to quote e.g. "IQR"}

\item{shownval}{boolean indicating if the n value used after na removal should be displayed in non fp cases. default is TRUE}

\item{total}{boolean indicating if a 'total' column should be added to the data frame. default is TRUE}

\item{rnd_digs}{the number of digits to round results to. default is 2}

\item{rownvar}{name of the first column which contains either indvar categories or function names. default is 'Class'}

\item{count_miss}{switch determining if NA values should be added to the indvar frequencies as it's own category. if this is 'ifmiss' then missing will be added. if this is 'none' then missing will be excluded and a message will show number of records removed. default is 'ifmiss'}

\item{count_miss_lab}{if count_miss='ifmiss' meaning we want NA values included this argument determines what they are labeled as. default is 'Missing'}
}
\description{
given an independent variable and a class variable creates a dataframe that is a cross tabulation of the two with values being 'out of' the class variable columns
}
\examples{
fp_msd_class2_function()

}
\keyword{collinear}
\keyword{correlation}
\keyword{fp_msd_class2}
\keyword{multicollinear}
\keyword{variables}
