% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{freqtbl_basic}
\alias{freqtbl_basic}
\title{freqtbl_basic function}
\usage{
freqtbl_basic(
  variable,
  name = "Category",
  total = FALSE,
  countna = TRUE,
  orderbycat = FALSE,
  reverseorder = TRUE
)
}
\arguments{
\item{variable}{the variable from which a frequency table will be created}

\item{name}{the name to appear in the table for the different categories. defaults to 'Category'}

\item{total}{boolean argument determining if a total should be added to the bottom of the table. defaults to FALSE}

\item{countna}{boolean argument determining if na's should be omitted (FALSE) or added as a frequency category (TRUE). defaults to TRUE}

\item{orderbycat}{boolean argument determining if the table sould be sorted by its categories (TRUE) instead of its frequencies (FALSE). defaulst to FALSE}

\item{reverseorder}{boolean argument determining if the table should be sorted from lowest to highest frequency (FALSE). defaults to TRUE}
}
\description{
creates frequency tables the way I like them. with freq, valid pct, and in descending order of frequency. this function allows you to change the name of the category columns, and whether or not totals and NAs are displayed
}
\examples{
freqtbl_basic_function()
}
\keyword{freqtbl}
\keyword{freqtbl_basic}
\keyword{frequency}
\keyword{table}
