% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{freqtbl_superbasic}
\alias{freqtbl_superbasic}
\title{freqtbl_superbasic function}
\usage{
freqtbl_superbasic(variable, name = "Category", total = FALSE, countna = TRUE)
}
\arguments{
\item{variable}{the variable from which a frequency table will be created}

\item{name}{the name to appear in the table for the different categories. defaults to 'Category'}

\item{total}{boolean argument determining if a total should be added to the bottom of the table. defaults to FALSE}

\item{countna}{boolean argument determining if na's should be omitted (FALSE) or added as a frequency category (TRUE). defaults to TRUE}
}
\description{
creates frequency tables the way I like them. with freq, valid pct, and in descending order of frequency. this function allows you to change the name of the category columns, and whether or not totals and NAs are displayed
}
\examples{
freqtbl_superbasic_function()
}
\keyword{freqtbl}
\keyword{freqtbl_superbasic}
\keyword{frequency}
\keyword{table}
