% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{gee_model_output}
\alias{gee_model_output}
\title{gee_model_output function}
\usage{
gee_model_output(
  model,
  staror = TRUE,
  starpval = TRUE,
  digs = 3,
  knit_cis = TRUE
)
}
\arguments{
\item{model}{the output from a logistic regression model e.g. lrmodel<-lm(outcome~vars,data,family='binomial')}

\item{staror}{boolean indicating whether to apply star characters to odds ratios. default is TRUE}

\item{starpval}{boolean indicating whether to apply star characters to p values. default is TRUE}

\item{digs}{the number of digits to round the values to. default is 3}

\item{knit_cis}{boolean indicating if the confidence intervals should be knit into an extra variable in the format "[lwr-upr]"}
}
\description{
creates a table of coefficients, std errors, z vals, pvals (with stars), ORs, and 95pct confints (only tested on LR models though should work with most others)
}
\examples{
gee_model_output()

}
\keyword{lm}
\keyword{logistic}
\keyword{lr_model_output}
\keyword{odds}
\keyword{or}
\keyword{ratio}
