% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{\%rine\%}
\alias{\%rine\%}
\title{match any function - EXACT}
\usage{
pattern \%rine\% list
}
\arguments{
\item{pattern}{the list/vector of patterns to match}

\item{list}{the list/vector that will be compared against for matches}
}
\description{
An extension of '%in%' that cycles through every element of a list/vector and evaluates if there is an exact match to a different list/vector
}
\examples{
c(1,2,3,4,5) \%rin\% c(5,6,7,8,9,10,11,12) == TRUE TRUE FALSE FALSE TRUE | c(5,6,7,8,9,10,11,12) \%rin\% c(1,2,3,4,5) TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
\%rin\%
}
\keyword{'%rine%'}
\keyword{any}
\keyword{in}
\keyword{list}
\keyword{match}
