% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{h.oneprop}
\alias{h.oneprop}
\title{h.oneprop function}
\usage{
h.oneprop(prop1, h)
}
\arguments{
\item{prop1}{the stable proportion that can be estimated or expected}

\item{h}{the variable to aggregate by (i.e. the classes)}
}
\description{
given one proportion and the effect size this function calculated the other proportion (i.e. so instead of having two proportions and calculating the effect size between the two)
}
\examples{
h.oneprop_function()
}
\keyword{analysis}
\keyword{chi}
\keyword{h.oneprop}
\keyword{power}
\keyword{proportion}
\keyword{pwr}
\keyword{sqaure}
