% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{hc_explore}
\alias{hc_explore}
\title{hc_explore function}
\usage{
hc_explore(
  data,
  cvars,
  mes = "rawdata",
  metric = "euclidean",
  method = "ward",
  diss = FALSE,
  maxclust = 15,
  bindata = FALSE
)
}
\arguments{
\item{data}{a dataframe containing only the identifier and  raw data to be clustered}

\item{cvars}{vector of names in order for all variables to be used in the clustering}

\item{mes}{vector of distance measures to use. default is distance measures from}

\item{metric}{metric to use in the agnes cluster analysis. default is 'euclidean'}

\item{method}{method of hierarchical clustering to use. default is 'ward'}

\item{diss}{whether dissimilarity matrix is being passed (TRUE) or raw data (FALSE). default is FALSE}

\item{maxclust}{maximum cluster solutions to produce. default is 15}

\item{bindata}{is the data being analyzed binary data (TRUE) or not (FALSE). default is FASLE}
}
\description{
massive function for exploratory hierachical cluster analysis. computes hierarchical, kmeans, and hierarchical with kmeans relocation cluster analyses, adds them to the working dataset and returns other working information
}
\examples{
hc_explore_function()
}
\keyword{cluster}
\keyword{distance}
\keyword{fit}
\keyword{hc_explore}
\keyword{inter-cluster}
\keyword{outlier}
\keyword{solution}
