% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{ICD}
\alias{ICD}
\title{ICD function}
\usage{
ICD(compgrp, method = "euclidean")
}
\arguments{
\item{compgrp}{the final cluster solution in the form of a mean aggregation of the variables used in the clustering.}

\item{method}{the method to be used in the ASED function. default is 'eucludean'}
}
\description{
computes average inter-cluster distance for a solution (i.e. average distance between the clusters)
}
\examples{
ICD_function()
}
\keyword{ICD}
\keyword{cluster}
\keyword{distance}
\keyword{inter-cluster}
\keyword{outlier}
