% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{ifneg}
\alias{ifneg}
\title{ifneg function}
\usage{
ifneg(mat, t = 1, f = 3)
}
\arguments{
\item{mat}{matrix to be used (should not already be transposed as the matrix does this-if the martix is in the orientation needed for a barplot without transposing you may need to transpose it when feeding into this function)}

\item{t}{the value to map the negative values to. defaults to 1}

\item{f}{the value to map the positive values to. defaults to 3}
}
\description{
checks if the transposed values of a matrix are negative and maps different values to it if they are or are not
}
\examples{
ifneg_function()
}
\keyword{barplot}
\keyword{check}
\keyword{ifneg}
