% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlandr.R, R/image analysis.R
\name{imgproc}
\alias{imgproc}
\title{imgproc function}
\usage{
imgproc(imgfilelist, dfile)

imgproc(imgfilelist, dfile)
}
\arguments{
\item{imgfilelist}{the list of image files (all must be the same size and the pixel ranges must correspond to the ranges selected from the comparison image)}

\item{dfile}{the datafile that contains the pixel ranges and avg, pixper, grx, gry, and tmat values for the comparison (typically blank) document/image}
}
\description{
searches a list of image files in the pixel ranges given and returns the difference in pixel darkness averages, dark pix proportion (pixper), x gradient, y gradient, and correlation between pixel range matrices (EXPERIMENTAL<NEEDS UPDATING)

searches a list of image files in the pixel ranges given and returns the difference in pixel darkness averages, dark pix proportion (pixper), x gradient, y gradient, and correlation between pixel range matrices (EXPERIMENTAL<NEEDS UPDATING)
}
\examples{
imgproc_function()
imgproc_function()
}
\keyword{image}
\keyword{imgproc}
\keyword{picture}
\keyword{processing}
\keyword{search}
