% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R, R/highlandr.R
\name{indxDF}
\alias{indxDF}
\title{indxDF function}
\usage{
indxDF(dat, cslist, mxc = length(cslist))

indxDF(dat, cslist, mxc = length(cslist))
}
\arguments{
\item{dat}{a dataframe of the raw data to be clustered}

\item{cslist}{cluster solution list is a list of different hierarchical cluster solutions to be evaluated for model fit}

\item{mxc}{maximum cluster solution to check. default is set to all of the solutions in 'cslist' but can be fewer if desired}
}
\description{
creates the proper dataframe format and inputs it into the clustrecs function (http://epub.wu.ac.at/1542/)

creates the proper dataframe format and inputs it into the clustrecs function (http://epub.wu.ac.at/1542/)
}
\examples{
indxDF_function()
indxDF_function()
}
\keyword{cluster}
\keyword{distance}
\keyword{fit}
\keyword{indxDF}
\keyword{inter-cluster}
\keyword{outlier}
\keyword{solution}
