% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{ischecked}
\alias{ischecked}
\title{is checked function}
\usage{
ischecked(
  x,
  pos = "Checked",
  neg = "Unchecked",
  ret_pos = TRUE,
  ret_neg = FALSE
)
}
\arguments{
\item{x}{vector containing a dichotomous 'checked' and 'unchecked' state}

\item{pos}{the positive or affirmative state. default is 'Checked' (case sensitive)}

\item{neg}{the negative state. default is 'Unchecked'}

\item{ret_pos}{what the function returns for a positive instance. default is TRUE}

\item{ret_neg}{what the function returns for a negative instance. default is FALSE}
}
\description{
this function is for converting common output of checkbox type variables to something else (e.g. default is converting the word 'Checked' to TRUE, and 'Unchecked' to FALSE)
}
\examples{
ischecked_function()
}
\keyword{binary}
\keyword{checked}
\keyword{choice}
\keyword{dichotomous}
\keyword{ischecked}
\keyword{multi}
\keyword{multiple}
