% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{Jaccard}
\alias{Jaccard}
\title{Jaccard function}
\usage{
Jaccard(x, y)
}
\arguments{
\item{x}{case 1 (the dataset columns in this case should contain only those used for matching)}

\item{y}{case 2 (the dataset columns in this case should contain only those used for matching)}
}
\description{
jaccard index (measure of agreement between two cases on suites of binary variables) M.11 / (M.11 + M.10 + M.01)
}
\examples{
Jaccard_function()
}
\keyword{Jaccard}
\keyword{binary}
\keyword{distance}
