% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{list_to_vecs}
\alias{list_to_vecs}
\title{list_to_vecs function}
\usage{
list_to_vecs(x)
}
\arguments{
\item{x}{list of lists which you would like to convert to a list of vectors}
}
\description{
wrapper for n2n this function takes a list or dataframe that may contain lists and converts the list elements to vectors. This is useful when using functions like pivot_wider which can often return lists of lists.
}
\examples{
list_to_vecs_function()

}
\keyword{convert}
\keyword{list}
\keyword{list_to_vecs}
\keyword{n2n}
\keyword{pivot_wider}
\keyword{vector}
