% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{lr_model_output}
\alias{lr_model_output}
\title{lr_model_output function}
\usage{
lr_model_output(model, staror = TRUE, starpval = TRUE, digs = 3)
}
\arguments{
\item{model}{the output from a logistic regression model e.g. lrmodel<-lm(outcome~vars,data,family='binomial')}

\item{staror}{boolean indicating whether to apply star characters to odds ratios. default is TRUE}

\item{starpval}{boolean indicating whether to apply star characters to p values. default is TRUE}

\item{digs}{the number of digits to round the values to. default is 3}
}
\description{
creates a table of coefficients, std errors, z vals, pvals (with stars), ORs, and 95% confints (only tested on LR models though should work with most others)
}
\examples{
lr_model_output_function()

}
\keyword{lm}
\keyword{logistic}
\keyword{lr_model_output}
\keyword{odds}
\keyword{ratio}
