% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{maxclustL}
\alias{maxclustL}
\title{maxclustL function}
\usage{
maxclustL(cindex)
}
\arguments{
\item{cindex}{a vector of fit index values representing cluster fit and requiring a max distance to the left solution as coming from the clustIndex function in the cclust package. Andreas Weingessel, Evgenia Dimitriadou and Sara Dolnicar, (An Examination Of Indexes For Determining The Number Of Clusters In Binary Data Sets, http://epub.wu.ac.at/1542/)}
}
\description{
computes and returns the number of the cluster solution that has the maximum difference to the cluster at the left side in an ascending cluster list of fit values
}
\examples{
maxclustL_function()
}
\keyword{cluster}
\keyword{distance}
\keyword{fit}
\keyword{inter-cluster}
\keyword{maxclustL}
\keyword{outlier}
\keyword{solution}
