% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{MCsmry}
\alias{MCsmry}
\title{MCsmry function}
\usage{
MCsmry(elementlist, variable, total = FALSE)
}
\arguments{
\item{elementlist}{the list of multiple choice elements in the variable to look for and count}

\item{variable}{the variable you wish to summarize}

\item{total}{boolean argument determining if a total should be added to the bottom of the table. defaults to FALSE}
}
\description{
multiple choice summary summarizes variables that have multiple choice options listed with a delimeter this function gives a table with the count & % of all cases containing each element in the list
}
\examples{
MCsmry_function()
}
\keyword{MCsmry}
\keyword{choice}
\keyword{freqtbl}
\keyword{frequency}
\keyword{multiple}
\keyword{table}
