% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{MCsmry_list}
\alias{MCsmry_list}
\title{MCsmry list function}
\usage{
MCsmry_list(elementlist, variable, total = FALSE, delimiter = ", ")

MCsmry_list(elementlist, variable, total = FALSE, delimiter = ", ")
}
\arguments{
\item{elementlist}{the list of multiple choice elements in the variable to look for and count}

\item{variable}{the variable you wish to summarize}

\item{total}{boolean argument determining if a total should be added to the bottom of the table. defaults to FALSE}

\item{delimiter}{the delimiter you to use to separate your data into a list format}
}
\description{
this function gives a table with the count & % of all cases containing each element in the list

multiple choice summary list summarizes variables that have multiple choice options listed with a delimeter this function gives a table with the count & % of all cases containing each element in the list. this function should be used over regular mcsmry when the delimiter is essential for distinguishing the list elements as is sometimes the case with multiple choice number values.
}
\examples{
MCsmry_list_function()
MCsmry_list_function()
}
\keyword{MCsmry}
\keyword{MCsmry_list}
\keyword{choice}
\keyword{freqtbl}
\keyword{frequency}
\keyword{list}
\keyword{multiple}
\keyword{table}
