% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{minclustSD}
\alias{minclustSD}
\title{minclustSD function}
\usage{
minclustSD(cindex)
}
\arguments{
\item{cindex}{a vector of fit index values representing cluster fit and requiring a minimum of the second differences solution as coming from the clustIndex function in the cclust package. Andreas Weingessel, Evgenia Dimitriadou and Sara Dolnicar, (An Examination Of Indexes For Determining The Number Of Clusters In Binary Data Sets, http://epub.wu.ac.at/1542/)}
}
\description{
computes and returns the number of the cluster solution that has the minimum of the second differences in an ascending cluster list of fit values
}
\examples{
minclustSD_function()
}
\keyword{cluster}
\keyword{distance}
\keyword{fit}
\keyword{inter-cluster}
\keyword{minclustSD}
\keyword{outlier}
\keyword{solution}
