% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R, R/highlandr.R
\name{mint}
\alias{mint}
\title{time conversion function}
\usage{
mint(x)

mint(x)
}
\arguments{
\item{x}{is the timestamp argument to be converted.}
}
\description{
This function turns a timestamp into a time variable in secs and turns it into hours.

This function turns a timestamp into a time variable in secs and turns it into hours.
}
\examples{
mint_function()
mint_function()
}
\keyword{mint}
\keyword{time}
