% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{n2n}
\alias{n2n}
\title{n2n function}
\usage{
n2n(x)
}
\arguments{
\item{x}{list of lists which you would like to convert to a list of vectors. this must be all lists, if a list element is instead a vector this function will cause errors.}
}
\description{
base function used to create list_to_vecs, recommend using that instead. This function takes a list comprised only of lists and converts the list elements to vectors. list_to_vecs takes any list/dataframe and converts any elements that are lists to vectors. This is useful when using functions like pivot_wider which can often return lists of lists.
}
\examples{
n2n_function()

}
\keyword{convert}
\keyword{list}
\keyword{list_to_vecs}
\keyword{n2n}
\keyword{pivot_wider}
\keyword{vector}
