% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{nequals}
\alias{nequals}
\title{nequals format function}
\usage{
nequals(x, cap = FALSE)
}
\arguments{
\item{x}{is the number to put into n equals format.}

\item{cap}{boolean, determining whether to capitalize 'n'. Default is FALSE meaning lower case 'n'.}
}
\description{
This function formats a number as "(n=x)". formats n equals, mimic n equals in VBA
}
\examples{
nequals_function()
}
\keyword{equals}
\keyword{format}
\keyword{n}
\keyword{nequals}
