% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{pair_them_shits}
\alias{pair_them_shits}
\title{pair_them_shits function}
\usage{
pair_them_shits(d, pairvars)
}
\arguments{
\item{d}{dataframe that you want limited to paired observations}

\item{pairvars}{vector of quoted variable names in the dataset to be evaluated for pairwise completion.}
}
\description{
Given a wide dataset and a vector of variable names this function performs pairwise deletion of records on the given variables. That is, it will return a subset of the given dataframe where there are no na's in any of the given variables thus leaving a dataset where only fully paired observations are present.
}
\examples{
pair_them_shits()

}
\keyword{pair}
\keyword{paired}
\keyword{pairwise}
\keyword{shits}
\keyword{them}
