% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R, R/highlandr.R
\name{pct}
\alias{pct}
\title{pct function}
\usage{
pct(x, digits = 2, format = "dec")

pct(x, digits = 2, format = "dec")
}
\arguments{
\item{x}{a vector of dichotomous values (0/1) in dec, num, or char format.}

\item{digits}{number of digits to round the percent to}

\item{format}{the format that the vector x is in. can be 'dec', 'num', or 'char'}
}
\description{
takes a dichotomous vector (0/1) and turns it into a percent value with '%' following

takes a dichotomous vector (0/1) and turns it into a percent value with '%' following
}
\examples{
pct_function()
pct_function()
}
\keyword{pct}
\keyword{percent}
