% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{pct_incr_time}
\alias{pct_incr_time}
\title{pct_incr_time function}
\usage{
pct_incr_time(init, yr, inc)
}
\arguments{
\item{init}{the initial or starting value}

\item{yr}{the number of years or iterations to run through}

\item{inc}{the percent increase or decrease formatted as a number with 1 returning the same value, less than 1 reducing the value, greated than one increasin the value.}
}
\description{
given one proportion and the effect size this function calculated the other proportion (i.e. so instead of having two proportions and calculating the effect size between the two)
}
\examples{
pct_incr_time_function()

}
\keyword{compound}
\keyword{increase}
\keyword{over}
\keyword{pct_incr_time}
\keyword{percent}
\keyword{time}
