% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{percent}
\alias{percent}
\title{percent conversion function}
\usage{
percent(x, digits = 2, format = "f", ...)
}
\arguments{
\item{x}{is the numeric argument to be converted}

\item{digits}{is the number of digits the number should be rounded to}

\item{format}{is the format that the character should be}

\item{...}{additional arguments to pass to format or paste0}
}
\description{
takes a decimal number and turns it into a character with specified digits and '%' following
}
\examples{
percent_function()
}
\keyword{character}
\keyword{cton}
\keyword{number}
