% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_data_column}
\alias{plot_data_column}
\title{plot_data_column function}
\usage{
plot_data_column(data, column)
}
\arguments{
\item{data}{a dataframe of the raw data to be clustered}

\item{column}{the column containing the data to be plotted}
}
\description{
plotting cluster fit for simple matching coefficient
}
\examples{
plot_data_column_function()
}
\keyword{cluster}
\keyword{distance}
\keyword{fit}
\keyword{inter-cluster}
\keyword{outlier}
\keyword{plot_data_column}
\keyword{solution}
