% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power analysis.R
\name{prop_diff_detectable}
\alias{prop_diff_detectable}
\title{prop_diff_detectable function}
\usage{
prop_diff_detectable(from, to, by, sig = 0.05, pwr = 0.8, prop1)
}
\arguments{
\item{from}{the lowest n value to check}

\item{to}{the highest n value to check}

\item{by}{the interval between the lowest and highest n values t check (e.g. from n=100 to n=500 by 50 will return a table with 9 rows)}

\item{sig}{the significance value to be used in the power analysis. the default for this is 0.05}

\item{pwr}{the expected statistical power for the power analysis. the default for this is 0.80}

\item{prop1}{the stable proportion that can be estimated or expected}
}
\description{
creates a table with sample size, effect size, proportion 2 and the minimum detectable change in proportion at the given levels in order to see how the values change and to plot
}
\examples{
prop_diff_detectable_function()
}
\keyword{analysis}
\keyword{chi}
\keyword{power}
\keyword{prop_diff_detectable}
\keyword{proportion}
\keyword{pwr}
\keyword{sample}
\keyword{size}
\keyword{sqaure}
