% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{pval_star}
\alias{pval_star}
\title{pval_star function}
\usage{
pval_star(
  pvals,
  starmat,
  vec_return = FALSE,
  correction = "none",
  digs = "none"
)
}
\arguments{
\item{pvals}{a vector of p values used to determine the demarcation to apply}

\item{starmat}{a vector of values to which demarcation (e.g. '*') will be applied. if this is not supplied then the demarcation will be applied to the pvalues supplied in pvals}

\item{vec_return}{boolean indcating whether to return a vector instaed of a data frame. default is FALSE}

\item{correction}{switch indicating whether a p.value correction for multiple comparisons should be applied. this uses stats::p.adjust and takes any of those types. default is 'none'}

\item{digs}{the number of digits to round the values to. default is 'none' indicating no rounding}
}
\description{
takes a vector of pvalues and values and adds star demarcation depending on the p values given
}
\examples{
pval_star_function()

}
\keyword{pval_star}
\keyword{pvalue}
\keyword{star}
