% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{rangelist}
\alias{rangelist}
\title{rangelist conversion function}
\usage{
rangelist(x, quotes = TRUE, delimiter = ", ")
}
\arguments{
\item{x}{is the vector containing the elements to be concatenated}

\item{quotes}{boolean variable determining if each element should be surrounded by quotes (TRUE). the default for this argument is TRUE}

\item{delimiter}{string argument of what the delimiter should be betweeen elements. the defualt for this agrument is ", "}
}
\description{
concatenates elements of a vector with options for quotes and delimiter. mimics the rangelist function that I created in vba
}
\examples{
rangelist_function()
}
\keyword{character}
\keyword{number}
\keyword{rangelist}
