% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{reassign}
\alias{reassign}
\title{reassign function}
\usage{
reassign(reassignobj, compgrp, method = "euclidean")
}
\arguments{
\item{reassignobj}{case to be reassigned to a cluster. this must take the form of a row with as many columns as there were variables used in the clustering- the variables must also come in the same order}

\item{compgrp}{the final cluster solution in the form of a mean aggregation of the variables used in the clustering. variables must be in the same amount and order as reassignobj}

\item{method}{the method to be used in the ASED function. default is 'eucludean'}
}
\description{
reassign function for reassigning outlier cases to clusters after a solution has been made using lowest average euclidean distance (ASED)
}
\examples{
reassign_function()
}
\keyword{cluster}
\keyword{outlier}
\keyword{reassign}
