% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{rename.variables}
\alias{rename.variables}
\title{rename variables function}
\usage{
rename.variables(data, newname, oldname = NULL)
}
\arguments{
\item{data}{dataframe containing names to convert}

\item{newname}{list of new names that you want for the data frame, if you do not want to change the name of every column then a vector of old names the same length as this should als be supplied}

\item{oldname}{a vector of old names the same length as newname which will do a 1:1 match for replacement. if left blank all column names will be relpace (i.e. newname should be a vector the name length as ncol(data))}
}
\description{
easier way to rename variables over base r. give a new name and old name (or a vector of new and old names) and this function will change them in the dataset
}
\examples{
rename.variables_function()
}
\keyword{change}
\keyword{name}
\keyword{rename}
\keyword{replace}
\keyword{variables}
