% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{round_if_num}
\alias{round_if_num}
\title{round_if_num function}
\usage{
round_if_num(x, digs = 3, except = NULL)
}
\arguments{
\item{x}{The dataset for which you want the numeric variables rounded}

\item{digs}{the number of digits to round the numeric variables to}

\item{except}{a vector of numeric variables to leave out of the rounding}
}
\description{
This function takes as it's input a dataframe and rounds all numeric variables to the number of digs supplied thus avoiding the error when applying 'round' to a dataframe that contains non-numeric variables
}
\examples{
round_if_num()
}
\keyword{is.numric}
\keyword{round}
\keyword{round_if}
\keyword{round_if_num}
