% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{row_to_var}
\alias{row_to_var}
\title{row_to_var function}
\usage{
row_to_var(df)
}
\arguments{
\item{df}{a dataframe for which you want to turn the first row into variable names}
}
\description{
deprecated but remains for functional use. given a data frame takes the first row of the data frame and makes it the variable names and then removes the first row. this is useful when using rbind or when an imported dataset doesn't account for headers
}
\examples{
row_to_var_function()
}
\keyword{row}
\keyword{row_to_var}
\keyword{to}
\keyword{variable}
