% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{rown_to_var}
\alias{rown_to_var}
\title{rown_to_var function}
\usage{
rown_to_var(df, keeprn = FALSE, makefirst = TRUE, varname = "terms")
}
\arguments{
\item{df}{a dataframe for which you want to turn row names into a variable}

\item{keeprn}{boolean indicating whether you want the rownames kept as rownames after they get added as a variable. default is FALSE}

\item{makefirst}{boolean indicating whether you want to make the rownames the first variable in the dataframe. default is TRUE}

\item{varname}{the name given to the new variable containing the rownames. default is 'terms'}
}
\description{
creates a variable from the row names, optionally makes it the first variable, optionally removes the row names, and names it what you want
}
\examples{
rown_to_var_function()
}
\keyword{name}
\keyword{row}
\keyword{rown_to_var}
\keyword{to}
\keyword{variable}
