% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{sf12_v2}
\alias{sf12_v2}
\title{sf12_v2 function}
\usage{
sf12_v2(x, revcode_sf2 = FALSE)
}
\arguments{
\item{x}{data frame containing the 12 item sf12 version 2 (must be the version 2)}

\item{revcode_sf2}{boolean indicating if sf12 variable sf2 should be reverse coded. it appears that it should be but was not in the sas file used in the above mentioned link. default is FALSE meaning the variable will NOT be recoded}
}
\description{
sf12 version 2 scoring function adapted from SAS code found here https://labs.dgsom.ucla.edu/hays/pages/programs_utilities
}
\examples{
sf12_v2_function()

}
\keyword{sf12}
\keyword{sf12_v2}
\keyword{sf12v2}
