% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{sig_val_auto}
\alias{sig_val_auto}
\title{sig_val_auto function}
\usage{
sig_val_auto(
  variables,
  crossvar,
  data,
  crossvartype,
  testtypes,
  correction_type = "none",
  sig = 0.05,
  stars = TRUE,
  digs = 3,
  fisher_dd = FALSE
)
}
\arguments{
\item{variables}{a vector of variable names for variables that you want to cross}

\item{crossvar}{the variable to cross against all in the 'variables' argument}

\item{data}{the data frame containing the variables in 'variables' and 'crossvar'}

\item{crossvartype}{the type of the cross variable. types can be 'b', 'f' which are binary, and factor respectively}

\item{testtypes}{a vector of types of the variables. types can be 'b', 'f', or 'c' which are binary, factor, and continuous respectively}

\item{correction_type}{type of p value correction to apply for multiple comparisons. default is 'none'. this uses the p.adjust from stats package so any adjustment name there will work here}

\item{sig}{the significance value to use. default is .05}

\item{stars}{boolean indicating whether to add stars to significance values. default is TRUE}

\item{digs}{number of digits to round values to. default is 3}
}
\description{
significance values creator with adjustment option testtypes and crossvar type can take the form (b,f,c) for binary, factor, or continuous
}
\examples{
sig_val_auto_function()

}
\keyword{crosstab}
\keyword{p}
\keyword{pvalue}
\keyword{sig_val_auto}
\keyword{significance}
\keyword{value}
