% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R, R/highlandr.R
\name{smc}
\alias{smc}
\title{smc function}
\usage{
smc(x, y)

smc(x, y)
}
\arguments{
\item{x}{case 1 (the dataset columns in this case should contain only those used for matching)}

\item{y}{case 2 (the dataset columns in this case should contain only those used for matching)}
}
\description{
simple matching coefficient index (measure of agreement between two cases on suites of binary variables) (M.11 + M.00) / (M.11 + M.10 + M.01 + M.00)

simple matching coefficient index (measure of agreement between two cases on suites of binary variables) (M.11 + M.00) / (M.11 + M.10 + M.01 + M.00)
}
\examples{
smc_function()
smc_function()
}
\keyword{binary}
\keyword{distance}
\keyword{matching}
\keyword{simple}
\keyword{smc}
