% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{smc_dist}
\alias{smc_dist}
\title{smc_dist function}
\usage{
smc_dist(x, y)
}
\arguments{
\item{x}{case 1 (the dataset columns in this case should contain only those used for matching)}

\item{y}{case 2 (the dataset columns in this case should contain only those used for matching)}
}
\description{
simple matching coefficient distance (measure of agreement between two cases on suites of binary variables) 1-((M.11 + M.00) / (M.11 + M.10 + M.01 + M.00))
}
\examples{
smc_function()
}
\keyword{binary}
\keyword{distance}
\keyword{matching}
\keyword{simple}
\keyword{smc}
\keyword{smc_dist}
