% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{sumisna}
\alias{sumisna}
\title{count na's in a vector}
\usage{
sumisna(x, funct = "is.na")
}
\arguments{
\item{x}{the vector containing the na values.}

\item{funct}{the function to use to count the na's, theoretically can be changed to a different function to count something other than na's}
}
\description{
This function counts the number of na's in a vector.
}
\examples{
sumisna_function()
}
\keyword{na}
\keyword{sum}
