% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{tween}
\alias{tween}
\title{tween function}
\usage{
tween(x, bot, top)
}
\arguments{
\item{x}{is the number to check}

\item{bot}{the lower (bottom) number to check against}

\item{top}{the higher (top) number to check against}
}
\description{
This function checks whether a number is betwee two other numbers
}
\examples{
tween_function()
}
\keyword{between}
\keyword{tween}
