% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{unadjusted_ors}
\alias{unadjusted_ors}
\title{unadjusted_ors function}
\usage{
unadjusted_ors(vars, outcome, x, digs = 3, staror = TRUE, starpval = TRUE)
}
\arguments{
\item{vars}{the vector of variable names to use for piecemeal comparison to the outcome variable}

\item{outcome}{the outcome variable to use in the unadjusted models}

\item{x}{the dataframe containing the outcome and covariate variables}

\item{digs}{option passed to the lr_model_output function. The number of digits to round the values to. default is 3}

\item{staror}{option passed to the lr_model_output function. Boolean indicating whether to apply star characters to odds ratios. default is TRUE}

\item{starpval}{option passed to the lr_model_output function. Boolean indicating whether to apply star characters to p values. default is TRUE}
}
\description{
Given a vector of variable names, and outcome variable, and a data frame, this function will produce the unadjusted odds ratios and model output by using the lr_model_output function and running each variable independently against the outcome
}
\examples{
unadjusted_ors_function()

}
\keyword{adjust}
\keyword{lm}
\keyword{logistic}
\keyword{lr_model_output}
\keyword{odds}
\keyword{or}
\keyword{ratio}
\keyword{unadjusted}
\keyword{unadjusted_ors}
