% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{update_timer}
\alias{update_timer}
\title{update_timer function}
\usage{
update_timer(start_time, cur_time = Sys.time(), cur_iter, total_iter)
}
\arguments{
\item{start_time}{A 'Sys.time()' object documenting the time the loop started}

\item{cur_time}{A 'Sys.time()' object documenting the current time. default is current system time}

\item{cur_iter}{The current iteration of the for loop}

\item{total_iter}{The total number of iterations of the loop}
}
\description{
A more advanced update timer for elapsed time in a for loop
}
\examples{
update_timer()

}
\keyword{progress}
\keyword{time}
\keyword{timer}
