% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{update_timer_simple}
\alias{update_timer_simple}
\title{update_timer_simple function}
\usage{
update_timer_simple(start_time)
}
\arguments{
\item{start_time}{A 'Sys.time()' object documenting the time the loop started}
}
\description{
A simple update timer for elapsed time in a for loop
}
\examples{
update_timer_simple()
}
\keyword{progress}
\keyword{time}
\keyword{timer}
