% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{varcompare}
\alias{varcompare}
\title{varcompare function}
\usage{
varcompare(d, vars, vartypes, force_pc = FALSE)
}
\arguments{
\item{d}{data frame containing all variables to be compared}

\item{vars}{the names of all the variables in the dataset}

\item{vartypes}{a vector of types of the variables. types can be 'd', 'n', or 'c' which are dichotomous, nominal, and continuous respectively}
}
\description{
create function to create a matrix that compares all variable types with quazi correlation measures to assess multicollinearity using the following criteria. vartypes must be "dochotomous" - d , "nominal" - n, or "continuous" - c and the permutation determines what is used to compare. the following are the correlation metric and significance types used and the associated permuations. #phi, x2 (dd); #cramer's v, x2 (dn,nd,nn); #point biserial, t (dc,cd); #omega2, anova (nc,cn); #pearson or spearman (depending on normality-shapirowilks), correlation sigs (cc)
}
\examples{
varcompare_function()

}
\keyword{collinear}
\keyword{correlation}
\keyword{multicollinear}
\keyword{varcompare}
\keyword{variables}
