% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlandr.R, R/summary.R
\name{varlist}
\alias{varlist}
\title{varlist function}
\usage{
varlist(x)

varlist(x)
}
\arguments{
\item{x}{the dataset you wish to summarize}
}
\description{
turn the list of variables into a presentable and readable data frame by breaking the list of variables across columns based on what the set is divisible by

turn the list of variables into a presentable and readable data frame by breaking the list of variables across columns based on what the set is divisible by
}
\examples{
varlist_function()
varlist_function()
}
\keyword{var}
\keyword{variable}
\keyword{varlist}
