% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{wssplot}
\alias{wssplot}
\title{wssplot function}
\usage{
wssplot(data, centers, nc = 15)
}
\arguments{
\item{data}{the data you wish to create the plot for. must be only the data that is clustered}

\item{centers}{the centers resulting from a hierarchical cluster analysis}

\item{nc}{number of clusters you want to take the plot out to. defaults to '15'}
}
\description{
scree plot for heirarchical clustering with built in kmeans relocation
}
\examples{
wssplot_function()
}
\keyword{cluster}
\keyword{hierarchical}
\keyword{plot}
\keyword{scree}
\keyword{wssplot}
