% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlandr.R, R/plots.R
\name{wssplot_km}
\alias{wssplot_km}
\title{wssplot_km function}
\usage{
wssplot_km(data, nc = 15, seed = 1234)

wssplot_km(data, nc = 15, seed = 1234)
}
\arguments{
\item{data}{the data you wish to create the plot for. must be only the data that is clustered}

\item{nc}{number of clusters you want to take the plot out to. defaults to '15'}

\item{seed}{lets you set the random seed for the kmeans algorithm for reproducability}
}
\description{
scree plot for kmeans clustering

scree plot for kmeans clustering
}
\examples{
wssplot_km_function()
wssplot_km_function()
}
\keyword{cluster}
\keyword{kmeans}
\keyword{plot}
\keyword{scree}
\keyword{wssplot}
