% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{x2bundle}
\alias{x2bundle}
\title{x2bundle function}
\usage{
x2bundle(d1, d2)
}
\arguments{
\item{d1}{variable 1}

\item{d2}{variable 2}
}
\description{
creates a crosstab of two variables and returns the crosstabs, the chi square value, the chi square significance, and the effect size
}
\examples{
x2bundle_function()
}
\keyword{chisquare}
\keyword{x2}
\keyword{x2bundle}
