% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlandr.R, R/summary.R
\name{x2multibundle}
\alias{x2multibundle}
\title{x2multibundle function}
\usage{
x2multibundle(indvar, data, depvar_vector = NULL)

x2multibundle(indvar, data, depvar_vector = NULL)
}
\arguments{
\item{indvar}{the single variable to check against all dependent variables}

\item{data}{the dataset from which indvar and depvar_vector variables come}

\item{depvar_vector}{vector of names of the dependent variables to check against}
}
\description{
creates a crosstab of two variables and returns the crosstabs, the chi square value, the chi square significance, and the effect size

creates a crosstab of two variables and returns the crosstabs, the chi square value, the chi square significance, and the effect size
}
\examples{
x2multibundle_function()
x2multibundle_function()
}
\keyword{chisquare}
\keyword{x2}
\keyword{x2multibundle}
