% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R, R/highlandr.R
\name{cton}
\alias{cton}
\title{cton conversion function}
\usage{
cton(x)

cton(x)
}
\arguments{
\item{x}{is the character argument to be converted.}
}
\description{
This function changes characters to numbers

This function changes characters to numbers
}
\examples{
cton_function()
cton_function()
}
\keyword{character}
\keyword{cton}
\keyword{number}
