% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{fp_msd_class2_msdmulti}
\alias{fp_msd_class2_msdmulti}
\title{fp_msd_class2_msdmulti function}
\usage{
fp_msd_class2_msdmulti(
  x,
  vars,
  varnames = c("Variables", "Function(s)"),
  classvar = NULL,
  funct1 = "mean",
  funct2 = "sd",
  shownval = TRUE,
  total = TRUE,
  rnd_digs = 2
)
}
\arguments{
\item{x}{dataframe containing the data to be summarized}

\item{vars}{vector of variable names for the varaibles that are to be summarized. all variables must be numeric and present in the data frame}

\item{varnames}{vector of names to switch the first and second columns of the dataframe to. default is "Variables" and "Function(s)" respectively}

\item{classvar}{optional class variable to cross with the independent variables. this is is typically presented as a slice (e.g. data$variable)}

\item{funct1}{the first function to use (i.e. outside the parentheses) if fp='FALSE'. default is 'mean'. if supplying different functions be sure to quote e.g. "IQR"}

\item{funct2}{the second function to use (i.e. inside the parentheses) if fp='FALSE'. default is 'sd'. if supplying different functions be sure to quote e.g. "IQR"}

\item{shownval}{boolean indicating if the n value used after na removal should be displayed in non fp cases. default is TRUE}

\item{total}{boolean indicating if a 'total' column should be added to the data frame. default is TRUE}

\item{rnd_digs}{the number of digits to round results to. default is 2}

\item{rownvar}{name of the first column which contains either indvar categories or function names. default is 'Class'}

\item{count_miss}{switch determining if NA values should be added to the indvar frequencies as it's own category. if this is 'ifmiss' then missing will be added. if this is 'none' then missing will be excluded and a message will show number of records removed. default is 'ifmiss'}

\item{count_miss_lab}{if count_miss='ifmiss' meaning we want NA values included this argument determines what they are labeled as. default is 'Missing'}
}
\description{
This function is a wrapper for fp_msd_class2 when there are multiple variables to analyze using functions outside of fp. Given a data frame and a vector of variable names this function will apply fp_msd_class2 to each one and rbind the results into a dataframe
}
\examples{
fp_msd_class2_msdmulti_function()

}
\keyword{collinear}
\keyword{correlation}
\keyword{fp_msd_class2_msdmulti}
\keyword{multicollinear}
\keyword{variables}
