% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcs.R
\name{or_pct}
\alias{or_pct}
\title{or_pct function}
\usage{
or_pct(x, digs = 0, knit = FALSE)
}
\arguments{
\item{x}{value to change}

\item{digs}{digits to round to}

\item{knit}{a boolean indicating if the value and the label should be knit together (e.g. "10pct increase" vs c("10pct","increase")). Default is FALSE (i.e. c("10pct","increase"))}
}
\description{
given an odds ratio this function will convert it into a percent change
}
\examples{
or_pct()

}
\keyword{odds}
\keyword{or}
\keyword{or_pct}
\keyword{pct}
\keyword{percent}
\keyword{ratio}
