% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{sig_val_auto}
\alias{sig_val_auto}
\title{sig_val_auto function}
\usage{
sig_val_auto(
  variables,
  crossvar,
  data,
  crossvartype,
  testtypes,
  reporttest = TRUE,
  correction_type = "none",
  sig = 0.05,
  stars = TRUE,
  digs = 3,
  force_fisher = FALSE,
  force_anova = FALSE,
  force_t = FALSE
)
}
\arguments{
\item{variables}{a vector of variable names for variables that you want to cross}

\item{crossvar}{the variable to cross against all in the 'variables' argument}

\item{data}{the data frame containing the variables in 'variables' and 'crossvar'}

\item{crossvartype}{the type of the cross variable. types can be 'b', 'f' which are binary, and factor respectively}

\item{testtypes}{a vector of types of the variables. types can be 'b', 'f', or 'c' which are binary, factor, and continuous respectively}

\item{reporttest}{boolean indicating if the type of test run should be added as a variable, default is yes}

\item{correction_type}{type of p value correction to apply for multiple comparisons. default is 'none'. this uses the p.adjust from stats package so any adjustment name there will work here}

\item{sig}{the significance value to use. default is .05}

\item{stars}{boolean indicating whether to add stars to significance values. default is TRUE}

\item{digs}{number of digits to round values to. default is 3}

\item{force_fisher}{boolean indicating if fisher's exact test should be used with two binary variables even if other assumptions are violated (e.g. <20pct expected values below 5)}

\item{force_anova}{boolean indicating if anova should be used with a factor and continuous variable even normality (shapiro wilk) is violated}

\item{force_t}{boolean indicating if t test should be used with a binary and continuous variable even normality (shapiro wilk) is violated}
}
\description{
significance values creator with adjustment option testtypes and crossvar type can take the form (b,f,c) for binary, factor, or continuous
}
\examples{
sig_val_auto()

}
\keyword{crosstab}
\keyword{p}
\keyword{pvalue}
\keyword{sig_val_auto}
\keyword{significance}
\keyword{value}
